// webapp/src/lib/api/model/createMemberRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Kakeibo API
 * マルチテナント家計簿システム API
 * OpenAPI spec version: 1.0.0
 */
import type { MemberRole } from "./memberRole";

export interface CreateMemberRequest {
  email: string;
  /** @maxLength 100 */
  displayName: string;
  role: MemberRole;
}